requirements.txt
discord.py
mysql-connector-python
SQLAlchemy
python-dotenv
requests
pillow
